requirements.txt
faiss-cpu
grog
langchain-groq
PyPDF2
langchain_google_genai
langchain
streamlit
langchain_community
python-dotenv
streamlit-extras
pypdf